+++ CopilotKit/examples/next-openai/src/app/components/vacation-list.tsx
@@ -15,7 +15,30 @@ export type Destination = {
   activities: string;
 };
 
+const document: DocumentPointer = {
+  id: "2",
+  name: "Travel Pet Peeves",
+  sourceApplication: "Google Docs",
+  iconImageUri: "/images/GoogleDocs.svg",
+  getContents: () => {
+    return [
+      "1. Crowded tourist spots",
+      "2. Expensive souvenirs",
+      "3. Uncomfortable airplane seats",
+      "4. Language barriers",
+      "5. Lost luggage",
+      "6. Jet lag",
+      "7. Long layovers",
+      "8. Unpredictable weather",
+      "9. Poor internet connection",
+      "10. Local cuisine not matching taste",
+    ].join("\n");
+  },
+} as DocumentPointer;
+
 export function VacationList() {
+  useMakeCopilotDocumentReadable(document);
+
   return (
     <div className="px-4 sm:px-6 lg:px-8 bg-slate-50 py-4">
       <div className="sm:flex sm:items-center">