+++ CopilotKit/packages/react-textarea/src/types/base/autosuggestions-bare-function.ts
@@ -1,3 +1,5 @@
+import { DocumentPointer } from "@copilotkit/react-core";
+
 export interface InsertionEditorState {
   textBeforeCursor: string;
   textAfterCursor: string;
@@ -15,6 +17,7 @@ export type AutosuggestionsBareFunction = (
 export type Generator_InsertionOrEditingSuggestion = (
   editorState: EditingEditorState,
   prompt: string,
+  documents: DocumentPointer[],
   abortSignal: AbortSignal
 ) => Promise<ReadableStream<string>>;